Changed from stupid long long's to uint64_t

+++ code/factorizer.c
@@ -2,13 +2,14 @@
 #include "factor_list.h"
 #include "factorizer.h"
 #include "threads_utils.h"
+#include <stdint.h>
 #include <stdlib.h>
 
 void *factorize(void *starting_state)
 {
     factorizer_starting_state_t *st = (factorizer_starting_state_t *) starting_state;
     
-    for (int i = st->start; i * i < st->to_fact->num; i += st->step)
+    for (uint64_t i = st->start; i * i < st->to_fact->num; i += st->step)
     {
         if (st->to_fact->num % i == 0)
         {
